requirements.txt
google-api-python-client
streamlit
python-dotenv
google-generativeai